Introdude global helper
