+++ src/ContextRegistry.php
@@ -2,14 +2,9 @@
 
 namespace Castor;
 
-use Psr\Log\LoggerInterface;
-
 /** @internal */
 class ContextRegistry
 {
-    private static Context $initialContext;
-    private static LoggerInterface $logger;
-
     /** @var array<string, ContextDescriptor> */
     private array $descriptors = [];
     private ?ContextDescriptor $default = null;
@@ -64,26 +59,5 @@ class ContextRegistry
     public function getNames(): array
     {
         return array_keys($this->descriptors);
-    }
-
-    public static function setInitialContext(Context $initialContext): void
-    {
-        self::$initialContext = $initialContext;
-    }
-
-    public static function getInitialContext(): Context
-    {
-        // We always need a default context, for example when using exec() in a context builder
-        return self::$initialContext ?? new Context();
-    }
-
-    public static function setLogger(LoggerInterface $logger): void
-    {
-        self::$logger = $logger;
-    }
-
-    public static function getLogger(): LoggerInterface
-    {
-        return self::$logger ?? throw new \LogicException('Logger not set yet.');
     }
 }
